Rename MCXA USB EHCI helpers
