+++ embassy-mcxa/src/usb/mod.rs
@@ -24,11 +24,11 @@
 //! - One transfer descriptor in flight per endpoint direction at a time.
 //!
 //! # Register access
-//! MMIO register blocks come from `nxp-pac`. The local [`registers`] module only
+//! MMIO register blocks come from `nxp-pac`. The local [`ehci`] module only
 //! contains EHCI DMA structures and bit masks that are not generated by the PAC.
 
 mod clock;
-mod registers;
+mod ehci;
 
 use core::cell::UnsafeCell;
 use core::future::poll_fn;
@@ -38,17 +38,17 @@ use core::task::Poll;
 
 pub use clock::PhyConfig;
 use cortex_m::asm;
-use embassy_hal_internal::{Peri, PeripheralType};
-use embassy_sync::waitqueue::AtomicWaker;
-use embassy_usb_driver::{
-    Direction, EndpointAddress, EndpointAllocError, EndpointError, EndpointInfo, EndpointType, Event, Unsupported,
-};
-use registers::{
+use ehci::{
     DEVICEADDR_USBADR_SHIFT, DEVICEADDR_USBADRA, DTD_BUFFER_PAGE_MASK, DTD_BUFFER_PAGE_SIZE, DTD_NEXT_TERMINATE,
     DTD_TOKEN_ACTIVE, DTD_TOKEN_ERROR_MASK, DTD_TOKEN_IOC, DTD_TOKEN_TOTAL_MASK, DTD_TOKEN_TOTAL_SHIFT, EPCTRL_RXE,
     EPCTRL_RXR, EPCTRL_RXS, EPCTRL_RXT_SHIFT, EPCTRL_TXE, EPCTRL_TXR, EPCTRL_TXS, EPCTRL_TXT_SHIFT, PORTSC1_PFSC,
     QH_CAP_IOS, QH_CAP_MAXLEN_SHIFT, QH_CAP_ZLT, QueueHead, TransferDescriptor, USBCMD_RS, USBCMD_RST, USBCMD_SUTW,
     USBMODE_CM_DEVICE, USBMODE_SLOM, USBSTS_PCI, USBSTS_SLI, USBSTS_UEI, USBSTS_UI, USBSTS_URI,
+};
+use embassy_hal_internal::{Peri, PeripheralType};
+use embassy_sync::waitqueue::AtomicWaker;
+use embassy_usb_driver::{
+    Direction, EndpointAddress, EndpointAllocError, EndpointError, EndpointInfo, EndpointType, Event, Unsupported,
 };
 
 use crate::interrupt;
